FEAT Use std::variant instead of mapbox::variant.
This was a good excuse to go back and revisit the code and see what has
changed now that C++17 has stablized.  Kinda nifty, but not a huge change
from the code we had with the mapbox.  I see where they're going for
C++20, but wow, is that language becoming bigger and more beastly every
iteration.